+++ Engine/UI/GUIElement.cpp
@@ -19,7 +19,7 @@ void GUIElement::Draw(Graphics & gfx)
 
 void GUIElement::Update(Mouse * _mouse, Keyboard * kbd)
 {
-	if (Box.Overlaps(_mouse->GetPosX(), _mouse->GetPosY())) {
+	if (Box.Overlaps((float)_mouse->GetPosX(), (float)_mouse->GetPosY())) {
 		Hover = true;
 	}
 	else {
@@ -48,17 +48,17 @@ void GUIElement::DrawBorder(Graphics& gfx)
 {
 	if (eBorder.isOn) {
 		for (int s = 0; s < eBorder.borderSize; s++) {
-			for (int i = Box.left; i < Box.right; i++) {
-				gfx.PutPixel(i + 1, Box.top+s, { 0,0,0 });
+			for (int i = (int)Box.left; i < Box.right; i++) {
+				gfx.PutPixel(i + 1, (int)Box.top+s, { 0,0,0 });
 			}
-			for (int i = Box.left; i < Box.right; i++) {
-				gfx.PutPixel(i, Box.bottom-s, { 0,0,0 });
+			for (int i = (int)Box.left; i < Box.right; i++) {
+				gfx.PutPixel(i, (int)Box.bottom-s, { 0,0,0 });
 			}
-			for (int i = Box.top; i < Box.bottom; i++) {
-				gfx.PutPixel(Box.left-s, i, { 0,0,0 });
+			for (int i = (int)Box.top; i < Box.bottom; i++) {
+				gfx.PutPixel((int)Box.left-s, i, { 0,0,0 });
 			}
-			for (int i = Box.top; i < Box.bottom; i++) {
-				gfx.PutPixel(Box.right+s, i + 1, { 0,0,0 });
+			for (int i = (int)Box.top; i < Box.bottom; i++) {
+				gfx.PutPixel((int)Box.right+s, i + 1, { 0,0,0 });
 			}
 		}
 	}
